Rename the Extract command to Extract Subthought
"Extract" said what the command does to the current thought but not what it
produces. The command slices the selected text out of the cursor thought and
files it as a child, so "Extract Subthought" names the result and lines the
command up with the other subthought commands it sits beside in the Creating
Thoughts group.

The rename runs all the way down: the user-facing label, the command id, the
action type and its action creator, the icon component, and the four files that
carry the name. Command ids are persisted — getUserToolbar reads a customized
toolbar out of the user's =toolbar thought as a list of ids — but Extract has
never been in TOOLBAR_DEFAULT_COMMANDS, so only a toolbar someone assembled by
hand can name it, and no migration ships with this.
