+++ docs/cursor-and-caret.md
@@ -39,7 +39,7 @@ The `selection.ts` module groups its functions roughly into:
 - **Reads:** `isActive()`, `isCollapsed()`, `isText()`, `isThought()`, `isNote()`, `isOnFirstLine()`, `isOnLastLine()`, `isStartOfElementNode()`, `isEndOfElementNode()`, `offset()`, `offsetThought()`, `offsetFromNode()`, `offsetStart()`, `offsetEnd()`, `offsetRange(editable)`, `text()`, `html()`, `getBoundingClientRect()`, `isNear(x, y, distance)`.
 - **Writes:** `set(node, { offset?, end? })`, `setRange(node, { start, end })`, `clear()`, `select(el)`, `removeCurrentSelection()`.
 - **Save/restore:** `save()` returns a `SavedSelection` opaque object; `restore(saved)` puts it back. Used when an action that re-renders the DOM needs to preserve the caret across the render.
-- **Split helpers:** `split(el)` and `splitNode(root, range)` return the HTML before/after the caret with formatting tags re-balanced. Used by the Split Sentences command and the Extract command.
+- **Split helpers:** `split(el)` and `splitNode(root, range)` return the HTML before/after the caret with formatting tags re-balanced. Used by the Split Sentences command and the Extract Subthought command.
 
 `set` always collapses the caret. To keep a *range* of text selected across a programmatic edit that re-renders the editable, capture it with `offsetRange(editable)` beforehand and restore it with `setRange(editable, range)` afterwards — both work in plain-text offsets, so nested formatting tags are handled for you. This is how the Letter Case command keeps the user's text selected ([issue #4840](https://github.com/cybersemics/em/issues/4840)). Offsets are only valid across an edit that preserves the text before them, so an edit that can change the length of the text has to map them: [`formatLetterCase`](../src/actions/formatLetterCase.ts) re-applies the letter case transform to the text preceding each offset, since `'ß'.toUpperCase()` is two characters. The restore has to wait for the re-render it is undoing, and a frame is not a reliable proxy for it: [`ContentEditable`](../src/components/ContentEditable.tsx) replaces the editable's contents from a passive effect, which can be flushed after the next animation frame, so a restore deferred with `requestAnimationFrame` can land on the old text and be wiped by the re-render, leaving nothing selected ([issue #4985](https://github.com/cybersemics/em/issues/4985)). `formatLetterCase` therefore observes the editable and restores the range on the mutation that produces the transformed text.
