+++ src/actions/index.ts
@@ -38,7 +38,7 @@ export { default as editThought } from './editThought'
 export { default as error } from './error'
 export { default as expandHoverDown } from './expandHoverDown'
 export { default as expandHoverUp } from './expandHoverUp'
-export { default as extractThought } from './extractThought'
+export { default as extractSubthought } from './extractSubthought'
 export { default as fontSize } from './fontSize'
 export { default as freeThoughts } from './freeThoughts'
 export { default as gestureMenu } from './gestureMenu'
