+++ src/commands/index.ts
@@ -26,7 +26,7 @@ export { default as delete } from './delete'
 export { default as deleteEmptyThoughtOrOutdent } from './deleteEmptyThoughtOrOutdent'
 export { default as devices } from './devices'
 export { default as exportContext } from './exportContext'
-export { default as extractThought } from './extractThought'
+export { default as extractSubthought } from './extractSubthought'
 export { default as favorite } from './favorite'
 export { default as fontSizeDown } from './fontSizeDown'
 export { default as fontSizeUp } from './fontSizeUp'
