+++ src/constants.ts
@@ -213,7 +213,7 @@ export const TOOLBAR_DEFAULT_COMMANDS: CommandId[] = [
   // 'cursorUp',
   // 'deleteEmptyThoughtOrOutdent',
   // 'deleteThoughtWithCursor',
-  // 'extractThought',
+  // 'extractSubthought',
   // 'help',
   // 'home',
   // 'join',
@@ -541,7 +541,7 @@ export const COMMAND_GROUPS: {
       'newSubthoughtTop',
       'newUncle',
       'newGrandChild',
-      'extractThought',
+      'extractSubthought',
       'generateThought',
     ],
   },
